Was having problems removing items in a list using list comprehension
ChatGPT helped me with this problem:
The problem is that i was modifying the list sortd while iterating
over it. When you remove an item from the list, the rest of the items
shift towards the beginning of the list, and the for loop does not
continue to the next item.

To fix this, you can iterate over a copy of the list using for
item in sortd[:]. This will allow you to modify the original list
without affecting the loop.